requirements.txt
asgiref==3.8.0
attrs==23.2.0
Django==4.2.10
djangorestframework==3.14.0
drf-spectacular==0.27.1
Faker==24.3.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
